Combined the two backdoor functions into one.
Syntax: backdoor(fileName, portNumber)

Where the portNumber is an optional argument. When left out,
the program will simply search for the file, like when detecting
the removal of the c99.php shell, for instance. When the port
number variable is there, it will make sure both the port is
closed, and the file is removed.